backend/requirements.txt
wheel
pyjwt
uvicorn
aiopath
aiofiles
pydantic
sqlalchemy
fastapi[all]
passlib[bcrypt]
